backend/requirements.txt
fastapi
numpy
opencv-python
pillow==7.2.0
python-multipart
uvicorn